% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_bar_chart.R
\name{R_bar_chart}
\alias{R_bar_chart}
\title{R bar chart}
\usage{
R_bar_chart(request, A2 = 0.483)
}
\description{
R bar chart
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesco Darino
}
