% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_CL.R
\name{request_CL}
\alias{request_CL}
\title{Create Cpk graph for a given request}
\usage{
request_CL(data, request, A2 = 0.483, d2 = 5.534)
}
\arguments{
\item{request}{\code{numeric} used to identify the rquest}

\item{A2}{\code{numeric} defines the boundaries of the control limits, default value 0.483}

\item{d2}{\code{numeric}}
}
\description{
TBD
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino
}
