% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_CL.R
\name{request_CL}
\alias{request_CL}
\title{Create Cpk graph for a given request}
\usage{
request_CL(data, request, A2 = 0.483, d2 = 5.534)
}
\arguments{
\item{data}{\code{data.frame} The dataset on which chart will be based.
For this function to work, your dataset must have :
a `Request` column that has a number that identifies the request,
columns starting with `Measure` that has the values for you measurements,
a `Process Sample` column that records the number of the sample per request,
a `Target value` column that gives us the the value on which we base the specifications.}

\item{request}{\code{numeric} Number used to identify the request}

\item{A2}{\code{numeric} constant defining the specification limits, default value 0.483}

\item{d2}{\code{numeric} constant defining the variability of the Process, default value of 5.534}
}
\description{
Displays a graph with a histogram with limits and text that show how much the outputs
of a process meet a given set specifications.
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino
}
