% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat.R
\name{summary_stat}
\alias{summary_stat}
\title{Summary table}
\usage{
summary_stat(data, request, A2 = 0.483)
}
\description{
Summary table
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesco Darino
}
